requirements.txt
Django~=3.0.7
Pillow~=7.1.2